init changed with user/password set as test/test

+++ init.bash
@@ -1,6 +1,12 @@
 cd seguinus
-cp -nv settings.py.example settings.py
+if [ ! -f settings.py ]; then
+    echo "copy settings"
+    cp -nv settings.py.example settings.py
+fi
 cd ..
-cp -nv preferences.py.example preferences.py 
+if [ ! -f preferences.py ]; then
+    echo "copy preferences"
+    cp -nv preferences.py.example preferences.py 
+fi
 python manage.py migrate
-python manage.py loaddata initData.json # TODO only execute this if there was no previous settings.py
+python manage.py loaddata initData.json
